requirements.txt
sympy==1.11.1
flask==2.2.3
matplotlib==3.7.1
numpy~=1.24.2
Pillow==9.5.0